// src/app/model/InputCard.ts
export class InputCard {
  // +
  userUuid: string;
  // +
  billName: string;
  billUuid: string;
  // +
  subcategoryName: string;
  subcategoryUuid: string;
  // +
  categoryName: string;
  categoryUuid: string;
  // +
  note: string;
  // +
  currency: string;
  // +
  sum: number;
  // +
  discount: number;
  // +
  count: number;
  // +
  unit: string;
  createCardDate: string;
}